pyproject.toml
[project]
name = "1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate==1.5.2",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.13",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "attrs==25.3.0",
    "audioread==3.0.1",
    "beautifulsoup4==4.13.3",
    "black==25.1.0",
    "bs4==0.0.2",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "datasets==3.4.0",
    "decorator==4.4.2",
    "dill==0.3.8",
    "distro==1.9.0",
    "dotenv==0.9.9",
    "filelock==3.13.1",
    "frozenlist==1.5.0",
    "fsspec==2024.6.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "huggingface-hub==0.29.3",
    "idna==3.10",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "jinja2==3.1.4",
    "jiter==0.9.0",
    "joblib==1.4.2",
    "lazy-loader==0.4",
    "librosa==0.11.0",
    "llvmlite==0.44.0",
    "markupsafe==2.1.5",
    "moviepy==1.0.3",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "multiprocess==0.70.16",
    "mypy-extensions==1.0.0",
    "networkx==3.3",
    "numba==0.61.0",
    "numpy==2.1.3",
    "openai==1.66.3",
    "packaging==24.2",
    "pandas==2.2.3",
    "pathspec==0.12.1",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "pooch==1.8.2",
    "proglog==0.1.10",
    "propcache==0.3.0",
    "psutil==7.0.0",
    "pyarrow==19.0.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "setuptools==70.2.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "soupsieve==2.6",
    "soxr==0.5.0.post1",
    "sympy==1.13.1",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "transformers==4.49.0",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "xxhash==3.5.0",
    "yarl==1.18.3",
]
